Emacs + Evil
This is seriously like my 4th or 5th time to try emacs. I'm making more
progress this time and my vim emulation is far closer to what I've
already got in my vimrc.

The approach this time is to use Cask for package management, evil for
vim emulation, and projectile and helm for other niceties.

The set up is following mostly what I've seen at
http://zeekat.nl/articles/making-emacs-work-for-me.html and
https://github.com/aaronbieber/dotfiles/.