feat: rework install to be config based
I saw dot (https://github.com/ubnt-intrepid/dot) and it piqued my interest; however, it didn't support fully some of the scenarios I wanted... this was just a minor tweak to my existing Rakefile. I like convention over configuration most of the time, but it was just odd in this case (e.g. XDG files, platform-specific, etc.) so this seemed better.

It is just a start.